refactor: move Name struct and service to separate module
